+++ desktop/cmp/grid/impl/filter/headerfilter/HeaderFilter.ts
@@ -12,6 +12,7 @@ import {button, buttonGroup} from '@xh/hoist/desktop/cmp/button';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
 import {Icon} from '@xh/hoist/icon';
+import {wait} from '@xh/hoist/promise';
 import {stopPropagation} from '@xh/hoist/utils/js';
 import {HeaderFilterModel} from './HeaderFilterModel';
 
@@ -22,7 +23,7 @@ import {HeaderFilterModel} from './HeaderFilterModel';
  */
 export const headerFilter = hoistCmp.factory({
     model: creates(HeaderFilterModel),
-    render() {
+    render({model}) {
         return panel({
             title: `Filter`,
             className: 'xh-column-header-filter',
@@ -31,7 +32,30 @@ export const headerFilter = hoistCmp.factory({
             onDoubleClick: stopPropagation,
             headerItems: [switcher()],
             item: tabContainer(),
-            bbar: bbar()
+            bbar: bbar(),
+            hotkeys: [
+                {
+                    allowInInput: true,
+                    combo: 'enter',
+                    label: 'Apply',
+                    group: 'Column Filter',
+                    onKeyDown: () =>
+                        // Wait for debounced reaction in `ValuesTabModel` to run before committing
+                        wait(400).then(
+                            () => (model.hasFilter || model.hasPendingFilter) && model.commit()
+                        )
+                },
+                {
+                    allowInInput: true,
+                    combo: 'escape',
+                    label: 'Cancel',
+                    group: 'Column Filter',
+                    onKeyDown: () => {
+                        model.shouldCommitOnClose = false;
+                        model.parent.close();
+                    }
+                }
+            ]
         });
     }
 });